refactor(vmm): split device manager
