+++ virt/arcbox-vmm/src/device/mod.rs
@@ -11,8 +11,6 @@
 use std::collections::HashMap;
 use std::sync::{Arc, Mutex, RwLock};
 
-#[cfg(test)]
-use arcbox_virtio::net::VirtioNetHeader;
 use arcbox_virtio::{DeviceStatus, QueueConfig, VirtioDevice};
 
 use crate::error::{Result, VmmError};
@@ -20,11 +18,18 @@ use crate::irq::{Irq, IrqChip};
 use crate::memory::MemoryManager;
 
 mod checksum;
+mod dispatch;
 mod mmio_state;
 pub(crate) mod net_worker;
+mod poll;
+mod tree;
+
+#[cfg(test)]
+mod tests;
 
 use checksum::finalize_virtio_net_checksum;
 pub use mmio_state::{MmioDevice, VirtioMmioState, virtio_mmio};
+pub use tree::DeviceTreeEntry;
 
 /// Device identifier.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
@@ -495,7 +500,7 @@ impl DeviceManager {
     /// Spawns the net-io worker thread if `device_id` is the primary VirtioNet
     /// and the worker has not already been spawned. Called from the DRIVER_OK
     /// handler (which only has `&self`).
-    fn maybe_spawn_net_rx_worker(
+    pub(super) fn maybe_spawn_net_rx_worker(
         &self,
         device_id: DeviceId,
         mmio_arc: &Arc<RwLock<VirtioMmioState>>,
@@ -838,591 +843,10 @@ impl DeviceManager {
         self.bridge_net_device_id
     }
 
-    /// Finds device by MMIO address.
-    #[must_use]
-    pub fn find_by_mmio(&self, addr: u64) -> Option<DeviceId> {
-        for (base, id) in &self.mmio_map {
-            if let Some(device) = self.devices.get(id) {
-                if addr >= *base && addr < *base + device.info.mmio_size {
-                    return Some(*id);
-                }
-            }
-        }
-        None
-    }
-
-    /// Handles MMIO read.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the read fails.
-    pub fn handle_mmio_read(&self, addr: u64, size: usize) -> Result<u64> {
-        let device_id = self
-            .find_by_mmio(addr)
-            .ok_or_else(|| VmmError::Device(format!("No device at MMIO address {addr:#x}")))?;
-
-        let device = self
-            .devices
-            .get(&device_id)
-            .ok_or_else(|| VmmError::Device(format!("Device {} not found", device_id.0)))?;
-
-        let base = device.info.mmio_base.unwrap_or(0);
-        let offset = addr - base;
-
-        if let Some(state) = &device.mmio_state {
-            let state = state
-                .read()
-                .map_err(|e| VmmError::Device(format!("Failed to lock device state: {e}")))?;
-
-            // Handle config space reads - forward to actual device
-            if offset >= virtio_mmio::regs::CONFIG {
-                let config_offset = offset - virtio_mmio::regs::CONFIG;
-                if let Some(virtio_dev) = &device.virtio_device {
-                    let dev = virtio_dev.lock().map_err(|e| {
-                        VmmError::Device(format!("Failed to lock virtio device: {e}"))
-                    })?;
-                    let mut data = vec![0u8; size];
-                    dev.read_config(config_offset, &mut data);
-                    tracing::trace!(
-                        "Config read: device={} offset={:#x} size={} data={:?}",
-                        device_id.0,
-                        config_offset,
-                        size,
-                        &data[..size.min(8)]
-                    );
-                    return Ok(match size {
-                        1 => u64::from(data[0]),
-                        2 => u64::from(u16::from_le_bytes([data[0], data[1]])),
-                        4 => u64::from(u32::from_le_bytes([data[0], data[1], data[2], data[3]])),
-                        8 => u64::from_le_bytes([
-                            data[0], data[1], data[2], data[3], data[4], data[5], data[6], data[7],
-                        ]),
-                        _ => 0,
-                    });
-                }
-                return Ok(0);
-            }
-
-            let value = state.read(offset);
-            let result = match size {
-                1 => u64::from(value as u8),
-                2 => u64::from(value as u16),
-                4 => u64::from(value),
-                _ => u64::from(value),
-            };
-
-            Ok(result)
-        } else {
-            Ok(0)
-        }
-    }
-
-    /// Handles MMIO write.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the write fails.
-    pub fn handle_mmio_write(&self, addr: u64, size: usize, value: u64) -> Result<()> {
-        let device_id = self
-            .find_by_mmio(addr)
-            .ok_or_else(|| VmmError::Device(format!("No device at MMIO address {addr:#x}")))?;
-
-        let device = self
-            .devices
-            .get(&device_id)
-            .ok_or_else(|| VmmError::Device(format!("Device {} not found", device_id.0)))?;
-
-        let base = device.info.mmio_base.unwrap_or(0);
-        let offset = addr - base;
-
-        if let Some(state) = &device.mmio_state {
-            let old_status = {
-                let s = state
-                    .read()
-                    .map_err(|e| VmmError::Device(format!("Failed to lock device state: {e}")))?;
-                s.status
-            };
-
-            // Handle config space writes - forward to actual device
-            if offset >= virtio_mmio::regs::CONFIG {
-                let config_offset = offset - virtio_mmio::regs::CONFIG;
-                if let Some(virtio_dev) = &device.virtio_device {
-                    let mut dev = virtio_dev.lock().map_err(|e| {
-                        VmmError::Device(format!("Failed to lock virtio device: {e}"))
-                    })?;
-                    let data: Vec<u8> = match size {
-                        1 => vec![value as u8],
-                        2 => (value as u16).to_le_bytes().to_vec(),
-                        4 => (value as u32).to_le_bytes().to_vec(),
-                        8 => value.to_le_bytes().to_vec(),
-                        _ => return Ok(()),
-                    };
-                    dev.write_config(config_offset, &data);
-                }
-                return Ok(());
-            }
-
-            let value32 = match size {
-                1 => value as u32 & 0xFF,
-                2 => value as u32 & 0xFFFF,
-                4 | 8 => value as u32,
-                _ => value as u32,
-            };
-
-            // Write to MMIO state
-            {
-                let mut state = state
-                    .write()
-                    .map_err(|e| VmmError::Device(format!("Failed to lock device state: {e}")))?;
-                state.write(offset, value32);
-            }
-
-            // Handle special cases after write
-            match offset {
-                virtio_mmio::regs::STATUS => {
-                    let new_status = value32 as u8;
-
-                    // Handle feature acknowledgment
-                    if new_status & DeviceStatus::FEATURES_OK != 0
-                        && old_status & DeviceStatus::FEATURES_OK == 0
-                    {
-                        if let Some(virtio_dev) = &device.virtio_device {
-                            let mmio_state = state.read().map_err(|e| {
-                                VmmError::Device(format!("Failed to lock device state: {e}"))
-                            })?;
-                            let mut dev = virtio_dev.lock().map_err(|e| {
-                                VmmError::Device(format!("Failed to lock virtio device: {e}"))
-                            })?;
-                            dev.ack_features(mmio_state.driver_features);
-                            tracing::debug!(
-                                "Device {} acknowledged features: {:#x}",
-                                device_id.0,
-                                mmio_state.driver_features
-                            );
-                        }
-                    }
-
-                    // Handle device activation
-                    if new_status & DeviceStatus::DRIVER_OK != 0
-                        && old_status & DeviceStatus::DRIVER_OK == 0
-                    {
-                        if let Some(virtio_dev) = &device.virtio_device {
-                            let mut dev = virtio_dev.lock().map_err(|e| {
-                                VmmError::Device(format!("Failed to lock virtio device: {e}"))
-                            })?;
-                            dev.activate().map_err(|e| {
-                                VmmError::Device(format!("Failed to activate device: {e}"))
-                            })?;
-                            tracing::info!("Device {} activated", device_id.0);
-                        }
-
-                        // Spawn the net-io worker for the primary VirtioNet device.
-                        self.maybe_spawn_net_rx_worker(device_id, state);
-                    }
-
-                    // Handle device reset
-                    if new_status == 0 {
-                        if let Some(virtio_dev) = &device.virtio_device {
-                            let mut dev = virtio_dev.lock().map_err(|e| {
-                                VmmError::Device(format!("Failed to lock virtio device: {e}"))
-                            })?;
-                            dev.reset();
-                            tracing::info!("Device {} reset", device_id.0);
-                        }
-                    }
-                }
-                virtio_mmio::regs::QUEUE_NOTIFY => {
-                    let queue_idx = value32 as u16;
-                    // Log vsock TX notifications at trace level (per-kick hot path).
-                    if device.info.device_type == DeviceType::VirtioVsock && queue_idx == 1 {
-                        tracing::trace!("QUEUE_NOTIFY: vsock TX queue 1 kicked by guest!",);
-                    }
-                    tracing::trace!(
-                        "QUEUE_NOTIFY: device {} ({:?}) queue {}",
-                        device_id.0,
-                        device.info.device_type,
-                        queue_idx,
-                    );
-
-                    if let Some(virtio_dev) = &device.virtio_device {
-                        // Build QueueConfig from current MMIO state for the
-                        // notified queue index.
-                        let qcfg = {
-                            let mmio_state = state.read().map_err(|e| {
-                                VmmError::Device(format!("Failed to lock state: {e}"))
-                            })?;
-                            let qi = queue_idx as usize;
-                            if qi < 8 {
-                                QueueConfig {
-                                    desc_addr: mmio_state.queue_desc[qi],
-                                    avail_addr: mmio_state.queue_driver[qi],
-                                    used_addr: mmio_state.queue_device[qi],
-                                    size: mmio_state.queue_num[qi],
-                                    ready: mmio_state.queue_ready[qi],
-                                    gpa_base: self.guest_ram_gpa,
-                                }
-                            } else {
-                                QueueConfig::default()
-                            }
-                        };
-
-                        if let (Some(ram_base), ram_size) =
-                            (self.guest_ram_base, self.guest_ram_size)
-                        {
-                            // Build a guest memory slice covering the guest RAM region.
-                            // The host pointer `ram_base` maps to GPA `guest_ram_gpa`.
-                            // All GPA-based indices must subtract `gpa_base` to obtain
-                            // the correct offset within this slice.
-                            //
-                            // SAFETY: `ram_base` is the host mapping returned by
-                            // Virtualization.framework and is valid for `ram_size` bytes.
-                            let gpa_base = self.guest_ram_gpa as usize;
-                            let guest_mem =
-                                unsafe { std::slice::from_raw_parts_mut(ram_base, ram_size) };
-
-                            // VirtioBlock async path: dispatch to worker thread
-                            // instead of blocking the vCPU with synchronous I/O.
-                            if device.info.device_type == DeviceType::VirtioBlock {
-                                let workers = self
-                                    .blk_workers
-                                    .lock()
-                                    .unwrap_or_else(std::sync::PoisonError::into_inner);
-                                if let Some(handle) = workers.get(&device_id) {
-                                    tracing::trace!(
-                                        "blk async dispatch for device {}",
-                                        device_id.0
-                                    );
-                                    handle.dispatch(guest_mem, &qcfg, queue_idx);
-                                }
-                            }
-                            // VirtioNet TX (queue 1): extract ethernet frames
-                            // from guest memory and write to the network host fd.
-                            // This bypasses the generic process_queue — the
-                            // concrete `VirtioNet` owns its fd + TX cursor via
-                            // `NetPort` and implements the hot path itself.
-                            else if device.info.device_type == DeviceType::VirtioNet
-                                && queue_idx == 1
-                                && (self.primary_net.is_some() || self.bridge_net.is_some())
-                            {
-                                let is_bridge = self
-                                    .bridge_net_device_id
-                                    .is_some_and(|bid| bid == device_id);
-                                let typed = if is_bridge {
-                                    self.bridge_net.as_ref()
-                                } else {
-                                    self.primary_net.as_ref()
-                                };
-                                let net_completions = match typed {
-                                    Some(arc) => arc
-                                        .lock()
-                                        .map(|d| {
-                                            d.drain_tx_queue(&qcfg, finalize_virtio_net_checksum)
-                                        })
-                                        .unwrap_or_default(),
-                                    None => Vec::new(),
-                                };
-                                let _ = guest_mem; // unused on this branch now
-
-                                if !net_completions.is_empty() {
-                                    // Update used ring for completed TX descriptors.
-                                    // Translate GPAs to slice offsets (checked).
-                                    let Some(used_off) =
-                                        (qcfg.used_addr as usize).checked_sub(gpa_base)
-                                    else {
-                                        tracing::warn!(
-                                            "invalid used GPA {:#x} below ram base {:#x}",
-                                            qcfg.used_addr,
-                                            gpa_base
-                                        );
-                                        return Ok(());
-                                    };
-                                    let q_size = qcfg.size as usize;
-                                    let used_idx_off = used_off + 2;
-                                    let mut used_idx = u16::from_le_bytes([
-                                        guest_mem[used_idx_off],
-                                        guest_mem[used_idx_off + 1],
-                                    ]);
-                                    for &(head, len) in &net_completions {
-                                        let entry =
-                                            used_off + 4 + ((used_idx as usize) % q_size) * 8;
-                                        if entry + 8 <= guest_mem.len() {
-                                            guest_mem[entry..entry + 4]
-                                                .copy_from_slice(&(head as u32).to_le_bytes());
-                                            guest_mem[entry + 4..entry + 8]
-                                                .copy_from_slice(&len.to_le_bytes());
-                                            used_idx = used_idx.wrapping_add(1);
-                                        }
-                                    }
-                                    std::sync::atomic::fence(std::sync::atomic::Ordering::Release);
-                                    guest_mem[used_idx_off..used_idx_off + 2]
-                                        .copy_from_slice(&used_idx.to_le_bytes());
-
-                                    // Write avail_event in the used ring to request
-                                    // kicks from the guest on future TX submissions.
-                                    // With VIRTIO_F_EVENT_IDX, the guest checks
-                                    // vring_need_event(avail_event, new, old) before
-                                    // kicking. Setting avail_event = current avail_idx
-                                    // ensures the guest kicks on the next submission.
-                                    if let Some(avail_off) =
-                                        (qcfg.avail_addr as usize).checked_sub(gpa_base)
-                                    {
-                                        let avail_idx = u16::from_le_bytes([
-                                            guest_mem[avail_off + 2],
-                                            guest_mem[avail_off + 3],
-                                        ]);
-                                        let avail_event_off = used_off + 4 + q_size * 8;
-                                        if avail_event_off + 2 <= guest_mem.len() {
-                                            guest_mem[avail_event_off..avail_event_off + 2]
-                                                .copy_from_slice(&avail_idx.to_le_bytes());
-                                        }
-                                    }
-
-                                    if let Some(_irq) = device.info.irq {
-                                        {
-                                            let mut s = state.write().map_err(|e| {
-                                                VmmError::Device(format!(
-                                                    "Failed to lock state: {e}"
-                                                ))
-                                            })?;
-                                            s.trigger_interrupt(virtio_mmio::INT_VRING);
-                                        }
-                                        self.sync_irq_level(device_id);
-                                    }
-                                }
-                            } else {
-                                // Generic process_queue for all other devices.
-                                let mut dev = virtio_dev.lock().map_err(|e| {
-                                    VmmError::Device(format!("Failed to lock device: {e}"))
-                                })?;
-                                // Log vsock TX processing at trace level (per-kick hot path).
-                                let is_vsock_tx = device.info.device_type
-                                    == DeviceType::VirtioVsock
-                                    && queue_idx == 1;
-                                match dev.process_queue(queue_idx, guest_mem, &qcfg) {
-                                    Ok(completions) if !completions.is_empty() => {
-                                        if is_vsock_tx {
-                                            tracing::trace!(
-                                                "Vsock QUEUE_NOTIFY TX: {} completions processed!",
-                                                completions.len(),
-                                            );
-                                        }
-                                        tracing::trace!(
-                                            "Device {} queue {} processed {} completions",
-                                            device_id.0,
-                                            queue_idx,
-                                            completions.len()
-                                        );
-                                        // Console TX completions don't need interrupts —
-                                        // the guest doesn't wait for host ACK on console output.
-                                        // Skipping avoids interrupt storms with level-triggered SPIs.
-                                        let skip_irq = device.info.device_type
-                                            == DeviceType::VirtioConsole
-                                            && queue_idx == 1;
-                                        if !skip_irq {
-                                            {
-                                                let mut s = state.write().map_err(|e| {
-                                                    VmmError::Device(format!(
-                                                        "Failed to lock state: {e}"
-                                                    ))
-                                                })?;
-                                                s.trigger_interrupt(virtio_mmio::INT_VRING);
-                                            }
-                                            self.sync_irq_level(device_id);
-                                        }
-                                    }
-                                    Ok(_) => {
-                                        if is_vsock_tx {
-                                            tracing::trace!(
-                                                "Vsock QUEUE_NOTIFY TX: kicked but 0 completions \
-                                                 (last_avail_idx_tx may already be current)",
-                                            );
-                                        }
-                                    }
-                                    Err(e) => {
-                                        tracing::warn!(
-                                            "Device {} queue {} error: {e}",
-                                            device_id.0,
-                                            queue_idx
-                                        );
-                                    }
-                                }
-                            } // end else (non-VirtioNet)
-                        } else {
-                            tracing::trace!(
-                                "Device {} queue {} notified but no guest memory set",
-                                device_id.0,
-                                queue_idx
-                            );
-                        }
-                    } else {
-                        tracing::trace!(
-                            "Device {} queue {} notified (no device impl)",
-                            device_id.0,
-                            queue_idx
-                        );
-                    }
-                }
-                virtio_mmio::regs::INTERRUPT_ACK => {
-                    // Sync the GIC SPI level with the updated interrupt_status.
-                    // If all bits are cleared, the SPI goes low; if bits remain
-                    // (from a concurrent completion), the SPI stays high.
-                    self.sync_irq_level(device_id);
-                }
-                _ => {}
-            }
-        }
-
-        Ok(())
-    }
-
     /// Returns an iterator over all devices.
     pub fn iter(&self) -> impl Iterator<Item = &DeviceInfo> {
         self.devices.values().map(|d| &d.info)
     }
-
-    /// Returns device tree entries for all `VirtIO` devices.
-    #[must_use]
-    pub fn device_tree_entries(&self) -> Vec<DeviceTreeEntry> {
-        // Sort by MMIO base address so the FDT node order is deterministic.
-        // Linux discovers virtio-mmio devices in FDT order, so the first
-        // virtio-blk node becomes vda, the second vdb, etc. Without sorting,
-        // HashMap iteration order is arbitrary and block device naming becomes
-        // non-deterministic (root=/dev/vda may point at the wrong disk).
-        let mut entries: Vec<DeviceTreeEntry> = self
-            .devices
-            .values()
-            .filter_map(|d| {
-                if let (Some(base), Some(irq)) = (d.info.mmio_base, d.info.irq) {
-                    Some(DeviceTreeEntry {
-                        compatible: "virtio,mmio".to_string(),
-                        reg_base: base,
-                        reg_size: d.info.mmio_size,
-                        irq,
-                    })
-                } else {
-                    None
-                }
-            })
-            .collect();
-        entries.sort_by_key(|e| e.reg_base);
-        entries
-    }
-
-    /// Polls the bridge (vmnet) host fd for inbound frames and injects
-    /// them into the bridge VirtioNet RX queue. Thin shim that reads the
-    /// device's current MMIO-transport queue configuration, hands it to
-    /// `VirtioNet::poll_rx`, and returns whether any frame was injected.
-    /// Caller fires the used-ring interrupt on `true`.
-    pub fn poll_bridge_rx(&self) -> bool {
-        let Some(bridge_arc) = self.bridge_net.as_ref() else {
-            return false;
-        };
-        let Some(bridge_id) = self.bridge_net_device_id else {
-            return false;
-        };
-        let Some(device) = self.devices.get(&bridge_id) else {
-            return false;
-        };
-        let Some(mmio_arc) = device.mmio_state.as_ref() else {
-            return false;
-        };
-
-        // Build a snapshot of the RX queue (idx 0) from MMIO state.
-        let rx_qcfg = {
-            let Ok(mmio) = mmio_arc.read() else {
-                return false;
-            };
-            let qi = 0usize;
-            if !mmio.queue_ready[qi] || mmio.queue_num[qi] == 0 {
-                return false;
-            }
-            QueueConfig {
-                desc_addr: mmio.queue_desc[qi],
-                avail_addr: mmio.queue_driver[qi],
-                used_addr: mmio.queue_device[qi],
-                size: mmio.queue_num[qi],
-                ready: true,
-                gpa_base: self.guest_ram_gpa,
-            }
-        };
-
-        let Ok(dev) = bridge_arc.lock() else {
-            return false;
-        };
-        dev.poll_rx(&rx_qcfg)
-    }
-
-    /// Called from the vCPU run loop during WFI (guest idle). Returns
-    /// true if any data was injected (caller should trigger interrupt).
-    /// Thin shim that builds RX + TX `QueueConfig` snapshots from the
-    /// vsock device's MMIO state and forwards to
-    /// `VirtioVsock::poll_rx_injection`. The 400-line body that was here
-    /// previously now lives on the device.
-    pub fn poll_vsock_rx(&self) -> bool {
-        let Some(vsock_arc) = self.vsock.as_ref() else {
-            return false;
-        };
-        let Some(device) = self
-            .devices
-            .values()
-            .find(|d| d.info.device_type == DeviceType::VirtioVsock)
-        else {
-            return false;
-        };
-        let Some(mmio_arc) = device.mmio_state.as_ref() else {
-            return false;
-        };
-
-        let (rx_qcfg, tx_qcfg) = {
-            let Ok(mmio) = mmio_arc.read() else {
-                return false;
-            };
-            let rxi = 0usize;
-            if !mmio.queue_ready[rxi] || mmio.queue_num[rxi] == 0 {
-                return false;
-            }
-            let rx = QueueConfig {
-                desc_addr: mmio.queue_desc[rxi],
-                avail_addr: mmio.queue_driver[rxi],
-                used_addr: mmio.queue_device[rxi],
-                size: mmio.queue_num[rxi],
-                ready: true,
-                gpa_base: self.guest_ram_gpa,
-            };
-            let txi = 1usize;
-            let tx = if mmio.queue_ready[txi] && mmio.queue_num[txi] > 0 {
-                Some(QueueConfig {
-                    desc_addr: mmio.queue_desc[txi],
-                    avail_addr: mmio.queue_driver[txi],
-                    used_addr: mmio.queue_device[txi],
-                    size: mmio.queue_num[txi],
-                    ready: true,
-                    gpa_base: self.guest_ram_gpa,
-                })
-            } else {
-                None
-            };
-            (rx, tx)
-        };
-
-        let Ok(mut dev) = vsock_arc.lock() else {
-            return false;
-        };
-        dev.poll_rx_injection(&rx_qcfg, tx_qcfg.as_ref())
-    }
-}
-
-/// Device tree entry for FDT generation.
-#[derive(Debug, Clone)]
-pub struct DeviceTreeEntry {
-    /// Compatible string.
-    pub compatible: String,
-    /// Register base address.
-    pub reg_base: u64,
-    /// Register region size.
-    pub reg_size: u64,
-    /// IRQ number.
-    pub irq: Irq,
 }
 
 impl Default for DeviceManager {
@@ -1444,127 +868,3 @@ const _: () = {
         assert_sync::<DeviceManager>();
     }
 };
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::net::Ipv4Addr;
-
-    use arcbox_net::ethernet::{TcpFrameParams, build_tcp_ack_frame, build_udp_ip_ethernet};
-    use arcbox_net::nat_engine::checksum::{tcp_checksum, udp_checksum};
-
-    #[test]
-    fn test_device_registration() {
-        let mut manager = DeviceManager::new();
-        let id = manager.register(DeviceType::Serial, "serial0").unwrap();
-
-        let info = manager.get(id);
-        assert!(info.is_some());
-        assert_eq!(info.unwrap().name, "serial0");
-    }
-
-    #[test]
-    fn test_virtio_mmio_state() {
-        let state = VirtioMmioState::new(2, 0x1234_5678);
-
-        assert_eq!(
-            state.read(virtio_mmio::regs::MAGIC),
-            virtio_mmio::MAGIC_VALUE
-        );
-        assert_eq!(state.read(virtio_mmio::regs::VERSION), virtio_mmio::VERSION);
-        assert_eq!(state.read(virtio_mmio::regs::DEVICE_ID), 2);
-    }
-
-    #[test]
-    fn test_virtio_mmio_features() {
-        let mut state = VirtioMmioState::new(2, 0xDEAD_BEEF_CAFE_BABE);
-
-        // Read low 32 bits
-        assert_eq!(state.read(virtio_mmio::regs::DEVICE_FEATURES), 0xCAFE_BABE);
-
-        // Select high 32 bits
-        state.write(virtio_mmio::regs::DEVICE_FEATURES_SEL, 1);
-        assert_eq!(state.read(virtio_mmio::regs::DEVICE_FEATURES), 0xDEAD_BEEF);
-    }
-
-    #[test]
-    fn test_finalize_virtio_net_checksum_repairs_ipv4_tcp_frame() {
-        let params = TcpFrameParams {
-            src_ip: Ipv4Addr::new(10, 0, 2, 2),
-            dst_ip: Ipv4Addr::new(198, 18, 30, 95),
-            src_port: 36402,
-            dst_port: 443,
-            seq: 1234,
-            ack: 0,
-            window: 64240,
-            src_mac: [0x52, 0x54, 0xAB, 0xFA, 0x2A, 0x70],
-            dst_mac: [0x02, 0xAB, 0xCD, 0x00, 0x00, 0x01],
-        };
-        let mut frame = build_tcp_ack_frame(&params);
-        let tcp_start = 14 + 20;
-        frame[tcp_start + 13] = 0x02;
-        frame[tcp_start + 16..tcp_start + 18].fill(0);
-
-        let header = VirtioNetHeader {
-            flags: VirtioNetHeader::FLAG_NEEDS_CSUM,
-            gso_type: VirtioNetHeader::GSO_NONE,
-            hdr_len: 0,
-            gso_size: 0,
-            csum_start: tcp_start as u16,
-            csum_offset: 16,
-            num_buffers: 1,
-        };
-        let mut packet_data = header.to_bytes().to_vec();
-        packet_data.extend_from_slice(&frame);
-
-        finalize_virtio_net_checksum(&mut packet_data);
-
-        let frame = &packet_data[VirtioNetHeader::SIZE..];
-        let stored = u16::from_be_bytes([frame[tcp_start + 16], frame[tcp_start + 17]]);
-        let mut tcp_segment = frame[tcp_start..].to_vec();
-        tcp_segment[16..18].fill(0);
-
-        assert_ne!(stored, 0);
-        assert_eq!(
-            stored,
-            tcp_checksum(params.src_ip.octets(), params.dst_ip.octets(), &tcp_segment)
-        );
-    }
-
-    #[test]
-    fn test_finalize_virtio_net_checksum_repairs_ipv4_udp_frame() {
-        let src_ip = Ipv4Addr::new(10, 0, 2, 2);
-        let dst_ip = Ipv4Addr::new(10, 0, 2, 1);
-        let payload = b"hello dns";
-        let src_mac = [0x52, 0x54, 0xAB, 0xFA, 0x2A, 0x70];
-        let dst_mac = [0x02, 0xAB, 0xCD, 0x00, 0x00, 0x01];
-        let mut frame = build_udp_ip_ethernet(src_ip, dst_ip, 49152, 53, payload, src_mac, dst_mac);
-        let udp_start = 14 + 20;
-        frame[udp_start + 6..udp_start + 8].fill(0);
-
-        let header = VirtioNetHeader {
-            flags: VirtioNetHeader::FLAG_NEEDS_CSUM,
-            gso_type: VirtioNetHeader::GSO_NONE,
-            hdr_len: 0,
-            gso_size: 0,
-            csum_start: udp_start as u16,
-            csum_offset: 6,
-            num_buffers: 1,
-        };
-        let mut packet_data = header.to_bytes().to_vec();
-        packet_data.extend_from_slice(&frame);
-
-        finalize_virtio_net_checksum(&mut packet_data);
-
-        let frame = &packet_data[VirtioNetHeader::SIZE..];
-        let stored = u16::from_be_bytes([frame[udp_start + 6], frame[udp_start + 7]]);
-        let mut udp_datagram = frame[udp_start..].to_vec();
-        udp_datagram[6..8].fill(0);
-
-        assert_ne!(stored, 0);
-        assert_eq!(
-            stored,
-            udp_checksum(src_ip.octets(), dst_ip.octets(), &udp_datagram)
-        );
-    }
-}
